New walls
There are known issues with the walls, such as the cobweb of polygons at the beginning of the atrium and the lack of mortar in the entrance hall. There have been small changes to the set dressing to give an idea of the directional look we talked about in our meeting with Aaron. We discussed what we would do with the puzzle elements in MM and the tutorial which I have yet to put in but know that they'll be in soon enough